Progress on Manifest Validation
- When ingesting/opening a new manifest binary file, try and frontload as many validations and range checks as we can
- Start simplifying the accessors to leverage this: hotpath should be less branchy, using asserts to catch *user* errors
- Standardize entry point indices as `uint32_t`: `uint16_t` was annoying, and we were also assuming this was a 1-based index. insanely brittle. horrible. why?

+++ src/compile/impl/SlangReflector.cpp
@@ -528,7 +528,7 @@ CookResult<RawVariant> SlangReflector::Reflect(LinkedVariant& linked_variant,
     RawVariant rawVariant;
     rawVariant.VariantSuffix = MakeAssignmentSuffix(descriptor.Canonical);
     rawVariant.VariantDescription = DescribeAssignment(descriptor.Canonical);
-    rawVariant.VariantIndex = static_cast<uint32_t>(descriptor.Index);
+    rawVariant.VariantIndex = descriptor.Index;
 
     // this step extracts the global bindings, which are put into the raw variant first
     // the global bindings are sorted by placement ordering before return from this fn
